backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
sqlalchemy>=2.0.0
# CV Pipeline
opencv-python>=4.9.0
ultralytics>=8.2.0       # YOLOv8 + ByteTracker
scikit-learn>=1.4.0      # K-Means team classifier
numpy>=1.26.0
# Optional: GPU acceleration (comment out if CPU only)
# torch>=2.2.0
# torchvision>=0.17.0
